requirements.txt
mnemonic
bip32utils
python-bitcoinrpc
pycryptodome